removing issue, moving to the appropriate folder
